Moved CMS models to data project, moved .template.json for multi-project template
