+++ src/BasicWeb.Data/Pages/StandardPage.cs
@@ -1,7 +1,7 @@
 using Piranha.AttributeBuilder;
 using Piranha.Models;
 
-namespace BasicWeb.Models
+namespace BasicWeb.Data.Pages
 {
     [PageType(Title = "Standard page")]
     public class StandardPage  : Page<StandardPage>
