+++ src/BasicWeb.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27428.2002
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28621.142
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BasicWeb", "BasicWeb\BasicWeb.csproj", "{172FE771-EB78-46C4-BBAA-51DB121B3A04}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BasicWeb.Data", "BasicWeb.Data\BasicWeb.Data.csproj", "{B5E0D2C7-4B9C-4B2F-A5D8-158C8A9B3A1E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{172FE771-EB78-46C4-BBAA-51DB121B3A04}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{172FE771-EB78-46C4-BBAA-51DB121B3A04}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{172FE771-EB78-46C4-BBAA-51DB121B3A04}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B5E0D2C7-4B9C-4B2F-A5D8-158C8A9B3A1E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B5E0D2C7-4B9C-4B2F-A5D8-158C8A9B3A1E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B5E0D2C7-4B9C-4B2F-A5D8-158C8A9B3A1E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B5E0D2C7-4B9C-4B2F-A5D8-158C8A9B3A1E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
