+++ src/BasicWeb/Views/Shared/_Layout.cshtml
@@ -10,7 +10,7 @@
 
     <title>@ViewBag.Title</title>
 
-    @if (Model is Piranha.Models.IMeta) {
+    @if (Model is IMeta) {
         <partial name="Partial/_Meta.cshtml" />
     }
