+++ src/BasicWeb/Views/_ViewImports.cshtml
@@ -4,6 +4,6 @@
 @using Piranha.Models
 @using Piranha.Extend.Blocks
 
-@using BasicWeb.Models
+@using BasicWeb.Data.Pages
 
 @inject Piranha.AspNetCore.Services.IApplicationService WebApp
